+++ pages/index.js
@@ -1,12 +1,9 @@
 import Head from 'next/head'
-import Image from 'next/image'
 import { Card } from 'components'
 import styles from 'styles/Home.module.scss'
 import { ROUTER_PATH } from 'constants/routerPath'
-import config from 'package.json';
 
 export default function Home() {
-  const { name, version } = config || {};
   return (
     <div className={styles.container}>
       <Head>
@@ -35,16 +32,6 @@ export default function Home() {
           <Card href={ROUTER_PATH.POST} title="POST" />
         </div>
       </main>
-
-      <footer className={styles.footer}>
-        <a rel="noopener noreferrer">
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-        <span className={styles.version}>{`${name}`.toUpperCase() + ` / v${version}`}</span>
-      </footer>
     </div>
   )
 }